Update README.md
The ultimate plan is to have code that will gather data from lichess using lichess' API. From there it will gather all of a specific users games for a given time control (rapid). From this point it'll be about reading the PGN files of the games, and having python say how often you win with white, how often with black, how often when you play a London or Sicilian, or whatever else. After that it'd be great to have a rolling average of the percent of wins shown graphically, and somehow factor your own and opponents rating into that as well.